# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AI Alpha API",
    "description": "API documentation for AI Alpha",
    "version": "2.1"
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-Key"
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
    "/admin/register": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Register a new admin with a specified role.",
        "description": "Register a new admin with a specified role.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Admin registration details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "role_id": {
                  "type": "integer"
                }
              },
              "required": [
                "email",
                "username",
                "password",
                "role_id"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Admin registered successfully"
          },
          "400": {
            "description": "Bad request (missing data or invalid role)"
          },
          "409": {
            "description": "Conflict (email or username already exists)"
          },
          "500": {
            "description": "Database error or unexpected error"
          }
        }
      }
    },
    "/admin/login": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Authenticate an admin.",
        "description": "Authenticate an admin.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Admin login details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": [
                "username",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "500": {
            "description": "Database error or unexpected error"
          }
        }
      }
    },
    "/admin/{admin_id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Retrieve admin information.",
        "description": "Retrieve admin information.",
        "parameters": [
          {
            "name": "admin_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Admin information retrieved successfully"
          },
          "404": {
            "description": "Admin not found"
          },
          "500": {
            "description": "Database error or unexpected error"
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Update admin information.",
        "description": "Update admin information.",
        "parameters": [
          {
            "name": "admin_id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Admin update details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin updated successfully"
          },
          "400": {
            "description": "Bad request (invalid data)"
          },
          "404": {
            "description": "Admin not found"
          },
          "500": {
            "description": "Database error or unexpected error"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete admin.",
        "description": "Delete admin.",
        "parameters": [
          {
            "name": "admin_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Admin deleted successfully"
          },
          "404": {
            "description": "Admin not found"
          },
          "500": {
            "description": "Database error or unexpected error"
          }
        }
      }
    },
...
  },
...
}